Replace Ack with the Silver Searcher (ag)
This is simply a 'swap one for the other' topic branch. If we
collectively believe ag to be better perhaps we will want to have
everyone default to it vs ack. It requires users to install ag before
the plugin will work but the same is true for Ack on certain machines as
well.

I'll issue a pull request for team review but suspect we'll probably
just leave the topic branch for now and let folks merge their own
version of what tools they like. You can also manually install via
vundle and remap the ack commands in your personal config as well. This
is what I have been doing but feel like others would benefit from this
being in the core distro.